+++ packages/css/src/viewTransition.ts
@@ -2,5 +2,5 @@ import { generateIdentifier } from './identifier';
 
 // createViewTransition is used for locally scoping CSS view transitions
 // For now it is mostly just an alias of generateIdentifier
-export const createViewTransition = (debugId?: string) =>
+export const createViewTransition = (debugId?: string): string =>
   generateIdentifier(debugId);