+++ packages/integration/src/processVanillaFile.ts
@@ -70,7 +70,7 @@ export async function processVanillaFile({
   outputCss = true,
   identOption = process.env.NODE_ENV === 'production' ? 'short' : 'debug',
   serializeVirtualCssPath,
-}: ProcessVanillaFileOptions) {
+}: ProcessVanillaFileOptions): Promise<string> {
   type Css = Parameters<Adapter['appendCss']>[0];
   type Composition = Parameters<Adapter['registerComposition']>[0];
 
@@ -330,7 +330,7 @@ export function serializeVanillaModule(
   cssImports: Array<string>,
   exports: Record<string, unknown>,
   unusedCompositionRegex: RegExp | null,
-) {
+): string {
   const functionSerializationImports = new Set<string>();
   const exportLookup = new Map(
     Object.entries(exports).map(([key, value]) => [