+++ packages/webpack-plugin/src/loader.ts
@@ -12,6 +12,9 @@ import {
 import type { LoaderContext } from './types';
 import { debug, formatResourcePath } from './logger';
 import { ChildCompiler } from './compiler';
+import { createRequire } from 'node:module';
+
+const require = createRequire(import.meta.url);
 
 const virtualFileLoader = require.resolve(
   path.join(
@@ -46,7 +49,7 @@ const defaultIdentifierOption = (
 ): IdentifierOption =>
   identifiers ?? (mode === 'production' ? 'short' : 'debug');
 
-export default function (this: LoaderContext, source: string) {
+export default function (this: LoaderContext, source: string): void {
   const { identifiers } = loaderUtils.getOptions(this) as InternalLoaderOptions;
 
   const { name } = getPackageInfo(this.rootContext);
@@ -68,7 +71,7 @@ export default function (this: LoaderContext, source: string) {
     });
 }
 
-export function pitch(this: LoaderContext) {
+export function pitch(this: LoaderContext): void {
   const { childCompiler, outputCss, identifiers, virtualLoader } =
     loaderUtils.getOptions(this) as InternalLoaderOptions;