+++ packages/private/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @vanilla-extract/private
 
+## 1.0.9
+
+### Patch Changes
+
+- [#1610](https://github.com/vanilla-extract-css/vanilla-extract/pull/1610) [`2b0be1b`](https://github.com/vanilla-extract-css/vanilla-extract/commit/2b0be1be15dcdc3df3248305fe80e566cce794cd) Thanks [@askoufis](https://github.com/askoufis)! - Revert "Improve ESM package entrypoints (#1597)" to fix `Named export not found` error when importing ESM entrypoints
+
+## 1.0.8
+
+### Patch Changes
+
+- [#1567](https://github.com/vanilla-extract-css/vanilla-extract/pull/1567) [`9d16de8`](https://github.com/vanilla-extract-css/vanilla-extract/commit/9d16de85b149b531f2e049e8d93bdb862667ff1f) Thanks [@andjsrk](https://github.com/andjsrk)! - Simplify `CSSVarFunction` type
+
+- [#1597](https://github.com/vanilla-extract-css/vanilla-extract/pull/1597) [`a7fccf8`](https://github.com/vanilla-extract-css/vanilla-extract/commit/a7fccf8a2626d610c060e095e0b9fb48a4ca5c9e) Thanks [@drwpow](https://github.com/drwpow)! - Fix ESM import path
+
+## 1.0.7
+
+### Patch Changes
+
+- [#1557](https://github.com/vanilla-extract-css/vanilla-extract/pull/1557) [`134117d`](https://github.com/vanilla-extract-css/vanilla-extract/commit/134117d32e25d92edc24c8c863d757294d7b5df8) Thanks [@askoufis](https://github.com/askoufis)! - Simplify `CSSVarFunction` type to improve type-checking performance
+
 ## 1.0.6
 
 ### Patch Changes